services/User/requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.0.1
idna==2.10
iniconfig==1.1.1
mysqlclient==2.0.3
packaging==20.9
pika==1.2.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5

# https://pypi.org/project/validate-docbr/
validate-docbr==1.8.2